pyproject.toml
[tool.poetry]
name = "rctf-client"
version = "0.1.0"
description = "CLI client for rCTF"
authors = ["Darin Mao <maodarin@gmail.com>"]

[tool.poetry.scripts]
rctf = "rctf_client.cli:rctf"

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.1.2"
requests = "^2.24.0"
pyyaml = "^5.3.1"
urwid = "^2.1.1"
pyperclip = "^1.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
